refactor: 📦 circular dependencies
