+++ src/lib/components/FlowDiagramExplorer/index.ts
@@ -1,2 +1,2 @@
 export type { FlowDiagramExplorerProps } from "./flow-diagram-explorer";
-export { default, DiagramConfigContext, defaultDiagramConfig } from "./flow-diagram-explorer";
+export { default } from "./flow-diagram-explorer";
