+++ src/lib/components/NodeActionHandler/index.ts
@@ -1,1 +1,2 @@
 export { NodeActionHandler } from "./node-action-handler";
+export { DiagramConfigContext, defaultDiagramConfig } from "./node-action-handler";
