+++ src/lib/render-library/index.ts
@@ -13,7 +13,6 @@ import { turbineRenderer } from "./Turbine";
 import { turbineGeneratorSetRenderer } from "./TurbineGeneratorSet";
 import { windTurbineRenderer } from "./WindTurbine";
 import { windTurbineSystemRenderer } from "./WindTurbineSystem";
-import { nodeRenderTypes } from "./node-render-types";
 
 export {
     cableRenderer,
@@ -31,5 +30,4 @@ export {
     turbineGeneratorSetRenderer,
     windTurbineRenderer,
     windTurbineSystemRenderer,
-    nodeRenderTypes,
 };
